add new components - totaltested, form and styled from them
