+++ src/assets/locales/deu/translation.json
@@ -26,7 +26,8 @@
     "Ваше_имя": "Ihr Name",
     "Возраст": "Alter",
     "Пол": "Geschlecht",
-    "Получить_результат": "Ergebnis abrufen"
+    "Получить_результат": "Ergebnis abrufen",
+    "Протестировано": "Totaler Test"
 
 
 }
