+++ src/assets/locales/en/translation.json
@@ -26,7 +26,8 @@
     "Ваше_имя": "Your name",
     "Возраст": "Age",
     "Пол": "Gender",
-    "Получить_результат": "Get result"
+    "Получить_результат": "Get result",
+    "Протестировано": "Total testing"
 
 
 }
