+++ src/assets/locales/ru/translation.json
@@ -25,5 +25,6 @@
     "Ваше_имя": "Ваше имя",
     "Возраст": "Возраст",
     "Пол": "Пол",
-    "Получить_результат": "Получить результат"
+    "Получить_результат": "Получить результат",
+    "Протестировано": "Всего протестирование"
 }
