+++ src/components/Header.jsx
@@ -5,6 +5,7 @@ import Logo from "./Logo";
 import { Outlet } from "react-router-dom";
 import { HeaderStyled } from "./styles/Header.Styled";
 import { StyledFlexWrapper } from "./styles/FlexWrapper.Styled";
+import TotalTested from "./TotalTested";
 
 const Header = () => {
   return (
@@ -12,8 +13,11 @@ const Header = () => {
       <HeaderStyled>
         <Container>
           <StyledFlexWrapper>
-            <Logo header/>
-            <Dropdown />
+            <Logo header />
+            <StyledFlexWrapper gap="48">
+              <TotalTested />
+              <Dropdown />
+            </StyledFlexWrapper>
           </StyledFlexWrapper>
         </Container>
       </HeaderStyled>
